convert some curl parameters to long explicitly

+++ IO/Network.cpp
@@ -99,7 +99,7 @@ namespace IO {
 
 				if (!multipart) {
 					if ((r = curl_easy_setopt(ch, CURLOPT_POSTFIELDS, gzip ? zip.getOutputPtr() : msg.c_str()))) throw r;
-					if ((r = curl_easy_setopt(ch, CURLOPT_POSTFIELDSIZE, gzip ? zip.getOutputLength() : msg.length()))) throw r;
+					if ((r = curl_easy_setopt(ch, CURLOPT_POSTFIELDSIZE, (long)(gzip ? zip.getOutputLength() : msg.length())))) throw r;
 				}
 				else if ((r = curl_easy_setopt(ch, CURLOPT_HTTPPOST, post)))
 					throw r;
@@ -108,9 +108,9 @@ namespace IO {
 				if ((r = curl_easy_setopt(ch, CURLOPT_HTTPHEADER, headers))) throw r;
 				if ((r = curl_easy_setopt(ch, CURLOPT_WRITEFUNCTION, curl_cb))) throw r;
 				if ((r = curl_easy_setopt(ch, CURLOPT_WRITEDATA, response))) throw r;
-				if ((r = curl_easy_setopt(ch, CURLOPT_NOPROGRESS, 1))) throw r;
+				if ((r = curl_easy_setopt(ch, CURLOPT_NOPROGRESS, (long)1))) throw r;
 				if (!userpwd.empty() && (r = curl_easy_setopt(ch, CURLOPT_USERPWD, userpwd.c_str()))) throw r;
-				if ((r = curl_easy_setopt(ch, CURLOPT_VERBOSE, 0))) throw r;
+				if ((r = curl_easy_setopt(ch, CURLOPT_VERBOSE, (long)0))) throw r;
 				if ((r = curl_easy_setopt(ch, CURLOPT_TIMEOUT, (long)TIMEOUT))) throw r;
 
 				if ((r = curl_easy_perform(ch))) throw r;
